requirements.txt
# Core Django
Django==5.1.3
psycopg2-binary==2.9.9
redis==4.5.4
celery[redis]==5.3.4

# Database for Render
dj-database-url==2.1.0

# Utilities
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
qrcode[pil]==7.4.2

# i18n
django-modeltranslation==0.19.7

# Static files  
whitenoise==6.7.0

# Production
gunicorn==22.0.0

# Development (only for local)
django-debug-toolbar==4.3.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
